refactoring + corrections on GraphVisualisation
