+++ src/api.js
@@ -1,4 +1,6 @@
-import { gql, GraphQLClient } from "graphql-request";
+import { GraphQLClient } from "graphql-request";
+import * as Base from "./api/Base";
+import * as BaseSepolia from "./api/BaseSepolia";
 
 // Hardcoded Endpoints with display names
 export const ENDPOINTS = {
@@ -9,236 +11,46 @@ export const ENDPOINTS = {
   baseSepolia: {
     url: "https://api.i7n.dev/v1/graphql",
     displayName: "Base Testnet",
+    module: BaseSepolia,
   },
   base: {
     url: "https://i7n.app/gql",
     displayName: "Base Mainnet",
+    module: Base,
   },
 };
 
 // Create GraphQL client based on endpoint
 export const createClient = (endpoint) => {
+  if (!ENDPOINTS[endpoint]) {
+    throw new Error(`Endpoint '${endpoint}' is not defined.`);
+  }
   return new GraphQLClient(ENDPOINTS[endpoint].url);
 };
 
+// Wrapper to dynamically select the appropriate module
+const getModuleForEndpoint = (endpoint) => {
+  if (!ENDPOINTS[endpoint] || !ENDPOINTS[endpoint].module) {
+    throw new Error(`No module defined for endpoint '${endpoint}'.`);
+  }
+  return ENDPOINTS[endpoint].module;
+};
+
+// Unified fetchTriples function
 export const fetchTriples = async (endpoint = "base") => {
-  const client = createClient(endpoint);
-  let query, data;
-
-  switch (endpoint) {
-    case "base":
-      query = gql`
-        query {
-          triples(limit: 1000) {
-            items {
-              id
-              subject {
-                label
-                id
-                creatorId
-                type
-              }
-              predicate {
-                label
-                id
-                creatorId
-                type
-              }
-              object {
-                label
-                id
-                creatorId
-                type
-              }
-            }
-          }
-        }
-      `;
-      data = await client.request(query);
-      return data.triples.items;
-    default:
-      query = gql`
-        query {
-          triples(limit: 1000) {
-            id
-            subject {
-              label
-              id
-              creatorId
-              type
-            }
-            predicate {
-              label
-              id
-              creatorId
-              type
-            }
-            object {
-              label
-              id
-              creatorId
-              type
-            }
-          }
-        }
-      `;
-      data = await client.request(query);
-      return data.triples;
-  }
+  const module = getModuleForEndpoint(endpoint);
+  return module.fetchTriples(endpoint);
 };
 
-// Fetch Atom Details
+// Unified fetchTriplesForNode function
+export const fetchTriplesForNode = async (nodeId, endpoint = "base") => {
+  const module = getModuleForEndpoint(endpoint);
+  debugger
+  return module.fetchTriplesForNode(nodeId, endpoint);
+};
+
+// Unified fetchAtomDetails function
 export const fetchAtomDetails = async (atomId, endpoint = "base") => {
-  const client = createClient(endpoint);
-  let query;
-  switch (endpoint) {
-    case "base":
-      query = gql`
-        query GetAtom($atomId: BigInt!) {
-          atom(id: $atomId) {
-            id
-            image
-            label
-            emoji
-            type
-            creatorId
-            vault {
-              totalShares
-            }
-          }
-        }
-      `;
-      break;
-    default:
-      query = gql`
-        query GetAtom($atomId: numeric!) {
-          atom(id: $atomId) {
-            id
-            image
-            label
-            emoji
-            type
-            creatorId
-            vault {
-              totalShares
-            }
-          }
-        }
-      `;
-  }
-
-  const variables = { atomId };
-
-  try {
-    const data = await client.request(query, variables);
-    return data.atom;
-  } catch (error) {
-    console.error("Error fetching atom details:", error);
-    throw error;
-  }
-};
-
-export const fetchTriplesForNode = async (nodeId, endpoint) => {
-  const client = createClient(endpoint);
-  let query, data, variables;
-
-  switch (endpoint) {
-    case "base":
-      query = gql`
-        query Triples($where: TripleFilter) {
-          triples(where: $where) {
-            items {
-                id
-                label
-                subject {
-                  label
-                  id
-                  creatorId
-                  type
-                }
-                predicate {
-                  label
-                  id
-                  creatorId
-                  type
-                }
-                object {
-                  label
-                  id
-                  creatorId
-                  type
-                }
-            }
-          }
-        }
-      `;
-      variables = {
-        where: {
-          OR: [
-            {
-              subjectId: parseInt(nodeId),
-            },
-            {
-              predicateId: parseInt(nodeId)
-            },
-            {
-              objectId: parseInt(nodeId)
-            },
-          ],
-        },
-      };
-      data = await client.request(query, variables);
-      debugger
-      return data.triples.items;
-    default:
-      query = gql`
-          query Triples($where: triples_bool_exp) {
-            triples(where: $where) {
-              id
-              label
-              subject {
-                label
-                id
-                creatorId
-                type
-              }
-              predicate {
-                label
-                id
-                creatorId
-                type
-              }
-              object {
-                label
-                id
-                creatorId
-                type
-              }
-            }
-          }
-      `;
-      variables = {
-        where: {
-          _or: [
-            {
-              predicateId: {
-                _eq: nodeId,
-              },
-            },
-            {
-              subjectId: {
-                _eq: nodeId,
-              },
-            },
-            {
-              creatorId: {
-                _eq: nodeId,
-              },
-            },
-          ],
-        },
-      };
-      data = await client.request(query, variables);
-      return data.triples;
-  }
+  const module = getModuleForEndpoint(endpoint);
+  return module.fetchAtomDetails(atomId, endpoint);
 };
